Compiler warnings on El Capitan.

+++ ext/libxml/ruby_xml_html_parser_context.c
@@ -232,7 +232,7 @@ static VALUE rxml_html_parser_context_string(VALUE klass, VALUE string)
     rb_raise(rb_eArgError, "Must specify a string with one or more characters");
 
   ctxt = xmlCreateMemoryParserCtxt(StringValuePtr(string),
-                                   RSTRING_LEN(string));
+                                   (int)RSTRING_LEN(string));
   if (!ctxt)
     rxml_raise(&xmlLastError);
